Moved files to the top level (#2)
This package is currently imported like import `go.rtnl.ai/commo/commo` because the files are in a folder, so I removed them from the folder so it's `import go.rtnl.ai/commo` like it should be. Also, some minor README and docs changes and updated a test to cover more configs.